refactor(Strings): move queryArgs extension
