+++ Blockchain/Extensions/Strings.swift
@@ -22,4 +22,21 @@ extension String {
     var urlQueryKeyPairDictionary: [String: String] {
         return [:]
     }
+
+    /// Returns query arguments from a string in URL format
+    var queryArgs: [String: String] {
+        var queryArgs = [String: String]()
+        let components = self.components(separatedBy: "&")
+        components.forEach {
+            let paramValueArray = $0.components(separatedBy: "=")
+
+            if paramValueArray.count == 2,
+                let param = paramValueArray[0].removingPercentEncoding,
+                let value = paramValueArray[1].removingPercentEncoding {
+                queryArgs[param] = value
+            }
+        }
+
+        return queryArgs
+    }
 }
